R code for analysis of 3 bowlers
Plots each game and series for 3 bowlers with rolling average since the linear regression model gave a terrible fit. Plots histograms, and dot plots for each game for a quick outlier check. Checks normality by shapiro-wilks, and performs wilcoxon- test on each bowler's game/series data, and compares games/series between bowlers to see if there is enough evidence to show that they are from different sets of data (in other words, whether one bowler is better or worse than another). Code could be optimized, and some sort of predictive analysis could be performed to see if the distributions will match in the future, but for now it does what I wanted it to for now. Note, one bowler missed the first week and had blind scores. Clearly this isn't reflective of their ability so that week was removed for bowler #2 scores. This must be remembered when picking files when inputting data.